Refactor pix2pose example to use in-library functions
